removed column/columns
This feature is not currently used and may never be used. This feature was a start towards allowing arbitrary
metadata on each column. It was optimized such that if two records had the same columns with the same metadata,
the metadata would only be stored once per chunk. Columns are still considered the "same" if their name is the same.
You were not allowed to have two columns with the same name and differing metadata.

This had a further simplication in that column info was stored in the XBD info block instead of in its own, separate
part of the chunk-file.

Not all of the above was implemented.

The main motifcation of all of this was noting if a specific field contained binary data or not.